+++ CHANGELOG.md
@@ -37,13 +37,14 @@
 
 **New**:
 
-- [SelectControl](src/controls/select.ts). Added option `multiple` (attribute \[multiple] & $options.multiple)
 - [SelectManyControl](src/controls/selectMany.ts) [**demo**](https://yegorich555.github.io/web-ui-pack/control/selectMany)
+- [SelectControl](src/controls/select.ts). Added option `multiple` (attribute \[multiple] & $options.multiple)
 - [PopupElement](src/popup/popupElement.ts).
   - Added properties `$isClose, $isOpening, $isClosing`
   - Calling `$show() & $hide()` possible several times at once without re-rendering
   - Events `$willShow & $willHide` have property `detail.showCase & detail.hideCase`
   - Added rule: popup can't be more than `window.innerWidth & window.innerHeight`. The rule impossible to disable!
+- [CircleElement](src/circleElement.ts). Added option `minsize` (attribute \[minsize] & $options.minsize)
 
 ## 0.4.1 (Jan 27, 2023)